Modified the BracketTree.py file to work with the project database.
Specifically, each node now contains information on the email, first name, last name, phone number, and address of each participant, read
directly from the database.

The uniquely-identifying email of the participant now serves as the "data" element of each node. Every other value is initialized to
"None."

Some of the functions and tests are edited.